requirements.txt
pydantic
openai
anthropic
langchain-openai

langchain 
chromadb 
tiktoken
langchain-text-splitters 
langchain-core 
langchain-community
langchain-chroma
langgraph

setuptools

# Free tier...
langchain-groq 
langchain-huggingface 
sentence-transformers
pypdf

python-dotenv

fastapi
python-multipart
uvicorn

requests
sqlalchemy
faiss-cpu
psycopg[binary]

# search package
duckduckgo-search
langchain-experimental
-e .
